requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
icalendar==6.3.1
idna==3.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
schedule==1.2.2
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
